+++ src/main/java/com/almostreliable/morejs/features/enchantment/PlayerEnchantEventJS.java
@@ -1,24 +1,44 @@
 package com.almostreliable.morejs.features.enchantment;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.EnchantmentInstance;
 import net.minecraft.world.level.Level;
 
-public class PlayerEnchantEventJS extends EnchantmentTableServerEventJS {
+import java.util.List;
 
-    private final int clickedButton;
+public class PlayerEnchantEventJS extends EnchantmentTableEventJS {
+    private final BlockPos pos;
+    private final int requiredLevel;
+    private final List<EnchantmentInstance> enchantments;
 
-    public PlayerEnchantEventJS(int clickedButton, ItemStack item, ItemStack secondItem, Level level, BlockPos pos, Player player, EnchantmentMenuState state) {
-        super(item, secondItem, level, pos, player, state);
-        this.clickedButton = clickedButton;
+    public PlayerEnchantEventJS(ItemStack item, ItemStack secondItem, Level level, BlockPos pos, Player player, EnchantmentMenuState state, int requiredLevel, List<EnchantmentInstance> enchantments) {
+        super(item, secondItem, level, player, state.getMenu());
+        this.pos = pos;
+        this.requiredLevel = requiredLevel;
+        this.enchantments = enchantments;
     }
 
-    public int getClickedButton() {
-        return clickedButton;
+    public List<EnchantmentInstance> getEnchantments() {
+        return enchantments;
     }
 
-    public EnchantmentData getSelected() {
-        return get(clickedButton);
+    public List<ResourceLocation> getEnchantmentIds() {
+        return getEnchantments()
+                .stream()
+                .flatMap(e -> e.enchantment.unwrapKey().stream())
+                .map(ResourceKey::location)
+                .toList();
+    }
+
+    public int getRequiredLevel() {
+        return requiredLevel;
+    }
+
+    public BlockPos getPosition() {
+        return pos;
     }
 }